requirements.txt
# ============================================================
# OpenClaw Agent Harness - Python 依赖
# ============================================================
# Python 版本要求: >= 3.10 (推荐 3.11/3.12)
# 创建环境: conda create -n agent-harness python=3.11 -y
# 安装依赖: pip install -r requirements.txt
# ============================================================

# --- LLM Provider ---
anthropic>=0.18.0          # Claude API
openai>=1.0.0             # OpenAI API
httpx>=0.25.0             # HTTP 客户端（用于 API 集成）

# --- 核心框架 ---
pyyaml>=6.0                # 配置文件解析

# --- Web UI (可选) ---
fastapi>=0.104.0          # 异步 Web 框架
uvicorn[standard]>=0.24.0 # ASGI 服务器
jinja2>=3.1.0             # 模板引擎

# --- GUI (可选) ---
PyQt6>=6.6.0              # 图形界面

# --- 数据处理 ---
pydantic>=2.0.0           # 数据验证
tiktoken>=0.5.0           # Token 计数
